pyproject.toml
[tool.poetry]
name = "flashcards"
version = "0.1.0"
description = ""
authors = ["AleksandraL"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
psycopg2 = "^2.9.9"
loguru = "^0.7.2"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
asyncpg = "^0.29.0"
argon2-cffi = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
